+++ include/colormanager.h
@@ -33,12 +33,12 @@ public:
 	void dump_config(std::vector<std::string>& config_output) const override;
 	void apply_colors(std::function<void(const std::string&, const std::string&)>
 		stfl_value_setter) const;
-	std::map<std::string, TextStyle> get_styles() const
-	{
-		return element_styles;
-	}
 
 private:
+	void emit_fallback_from_to(const std::string& from_element, const std::string& to_element,
+		const std::function<void(const std::string&, const std::string&)>& stfl_value_setter)
+	const;
+
 	std::map<std::string, TextStyle> element_styles;
 };