+++ include/pbview.h
@@ -2,10 +2,13 @@
 #define PODBOAT_VIEW_H_
 
 #include "colormanager.h"
-#include "keymap.h"
 #include "listwidget.h"
 #include "textviewwidget.h"
 #include "stflpp.h"
+
+namespace newsboat {
+class KeyMap;
+}
 
 using namespace newsboat;
 
@@ -13,8 +16,6 @@ namespace podboat {
 
 class PbController;
 class Download;
-
-struct KeyMapHintEntry;
 
 class PbView {
 public:
@@ -32,18 +33,12 @@ public:
 	}
 
 private:
-	struct KeyMapHintEntry {
-		Operation op;
-		char* text;
-	};
-
 	void run_help();
 	void set_dllist_keymap_hint();
 	void set_help_keymap_hint();
 	std::pair<double, std::string> get_speed_human_readable(double kbps);
 	void handle_resize();
 
-	std::string prepare_keymaphint(KeyMapHintEntry* hints);
 	std::string format_line(const std::string& podlist_format,
 		const Download& dl,
 		unsigned int pos,