+++ src/filebrowserformaction.cpp
@@ -300,11 +300,10 @@ void FileBrowserFormAction::init()
 	set_value("filenametext_pos", std::to_string(default_filename.length()));
 }
 
-KeyMapHintEntry* FileBrowserFormAction::get_keymap_hint()
+const std::vector<KeyMapHintEntry>& FileBrowserFormAction::get_keymap_hint() const
 {
-	static KeyMapHintEntry hints[] = {{OP_QUIT, _("Cancel")},
-		{OP_OPEN, _("Save")},
-		{OP_NIL, nullptr}
+	static const std::vector<KeyMapHintEntry> hints = {{OP_QUIT, _("Cancel")},
+		{OP_OPEN, _("Save")}
 	};
 	return hints;
 }