+++ test/colormanager.cpp
@@ -11,47 +11,87 @@
 
 using namespace newsboat;
 
+class StylesCollector {
+	std::map<std::string, std::string> styles;
+
+public:
+	StylesCollector() = default;
+
+	std::function<void(const std::string&, const std::string&)> setter()
+	{
+		return [this](const std::string& element, const std::string& style) {
+			if (this->styles.find(element) != this->styles.cend()) {
+				throw std::invalid_argument(std::string("Multiple styles for element ") + element);
+			}
+
+			this->styles[element] = style;
+		};
+	}
+
+	size_t styles_count() const
+	{
+		return styles.size();
+	}
+
+	std::string style(const std::string& element) const
+	{
+		const auto style = styles.find(element);
+		if (style != styles.cend()) {
+			return style->second;
+		} else {
+			return {};
+		}
+	}
+};
+
 TEST_CASE(
-	"get_styles() returns foreground/background colors and attributes for each "
-	"element that was processed",
+	"apply_colors() invokes the callback for each element, supplying the element name and its style",
 	"[ColorManager]")
 {
 	ColorManager c;
+	StylesCollector collector;
 
-	{
-		INFO("By default, the list is empty");
-		REQUIRE(c.get_styles().size() == 0);
+	SECTION("By default, the list is empty") {
+		c.apply_colors(collector.setter());
+		REQUIRE(collector.styles_count() == 0);
 	}
 
-	{
-		INFO("Each processed action adds corresponding entry to return "
-			"value");
-
+	SECTION("Each processed action adds corresponding entry to return value") {
 		c.handle_action("color", {"listnormal", "default", "default"});
-		REQUIRE(c.get_styles().size() == 1);
-		REQUIRE(c.get_styles().count("listnormal") == 1);
-
-		c.handle_action("color", {"listfocus", "cyan", "default", "bold", "underline"});
-		REQUIRE(c.get_styles().size() == 2);
-		REQUIRE(c.get_styles().count("listfocus") == 1);
-
+		c.handle_action("color", {"listfocus_unread", "cyan", "default", "bold", "underline"});
 		c.handle_action("color", {"background", "red", "yellow"});
-		REQUIRE(c.get_styles().size() == 3);
-		REQUIRE(c.get_styles().count("background") == 1);
+		c.handle_action("color", {"info", "green", "white", "reverse"});
+		c.handle_action("color", {"end-of-text-marker", "color123", "default", "dim", "protect"});
+		c.handle_action("color", {"hint-key", "default", "color2"});
+		c.handle_action("color", {"hint-description", "color3", "default"});
 
-		REQUIRE(c.get_styles()["listnormal"].fg_color == "default");
-		REQUIRE(c.get_styles()["listnormal"].bg_color == "default");
-		REQUIRE(c.get_styles()["listnormal"].attributes.size() == 0);
+		c.apply_colors(collector.setter());
 
-		REQUIRE(c.get_styles()["listfocus"].fg_color == "cyan");
-		REQUIRE(c.get_styles()["listfocus"].bg_color == "default");
-		REQUIRE(c.get_styles()["listfocus"].attributes.size() == 2);
-		REQUIRE(c.get_styles()["listfocus"].attributes[0] == "bold");
-		REQUIRE(c.get_styles()["listfocus"].attributes[1] == "underline");
+		REQUIRE(collector.styles_count() == 10);
+		REQUIRE(collector.style("listnormal") == "");
+		REQUIRE(collector.style("listfocus_unread") == "fg=cyan,attr=bold,attr=underline");
+		REQUIRE(collector.style("background") == "fg=red,bg=yellow");
+		REQUIRE(collector.style("info") == "fg=green,bg=white,attr=reverse");
+		REQUIRE(collector.style("title") == "fg=green,bg=white,attr=reverse");
+		REQUIRE(collector.style("end-of-text-marker") == "fg=color123,attr=dim,attr=protect");
+		REQUIRE(collector.style("hint-key") == "bg=color2");
+		REQUIRE(collector.style("hint-description") == "fg=color3");
 
-		REQUIRE(c.get_styles()["background"].fg_color == "red");
-		REQUIRE(c.get_styles()["background"].bg_color == "yellow");
-		REQUIRE(c.get_styles()["background"].attributes.size() == 0);
+		// These two weren't set explicitly and fell back to `info`
+		REQUIRE(collector.style("hint-keys-delimiter") == "fg=green,bg=white,attr=reverse");
+		REQUIRE(collector.style("hint-separator") == "fg=green,bg=white,attr=reverse");
+	}
+
+	SECTION("For `article` element, two additional elements are emitted") {
+		c.handle_action("color", {"article", "white", "blue", "reverse"});
+
+		c.apply_colors(collector.setter());
+
+		REQUIRE(collector.styles_count() == 3);
+		REQUIRE(collector.style("article") == "fg=white,bg=blue,attr=reverse");
+		REQUIRE(collector.style("color_bold") == "fg=white,bg=blue,attr=reverse,attr=bold");
+		REQUIRE(collector.style("color_underline") ==
+			"fg=white,bg=blue,attr=reverse,attr=underline");
 	}
 }
 
@@ -61,13 +101,17 @@ TEST_CASE("register_commands() registers ColorManager with ConfigParser",
 	ConfigParser cfg;
 	ColorManager clr;
 
+	StylesCollector collector;
+
 	REQUIRE_NOTHROW(clr.register_commands(cfg));
 
-	REQUIRE(clr.get_styles().size() == 0);
+	clr.apply_colors(collector.setter());
+	REQUIRE(collector.styles_count() == 0);
 
 	cfg.parse_file("data/config-with-colors");
 
-	REQUIRE(clr.get_styles().size() == 2);
+	clr.apply_colors(collector.setter());
+	REQUIRE(collector.styles_count() == 2);
 }
 
 TEST_CASE(
@@ -201,18 +245,98 @@ TEST_CASE("dump_config() returns everything we put into ColorManager",
 	REQUIRE(config.size() == 1);
 	REQUIRE(equivalent());
 
-	expected.emplace("color background green cyan bold");
-	c.handle_action("color", {"background", "green", "cyan", "bold"});
+	expected.emplace("color article green cyan bold");
+	c.handle_action("color", {"article", "green", "cyan", "bold"});
 	config.clear();
 	c.dump_config(config);
 	REQUIRE(config.size() == 2);
 	REQUIRE(equivalent());
 
-	expected.emplace("color listnormal black yellow underline standout");
+	expected.emplace("color listnormal_unread black yellow underline standout");
 	c.handle_action("color",
-	{"listnormal", "black", "yellow", "underline", "standout"});
+	{"listnormal_unread", "black", "yellow", "underline", "standout"});
 	config.clear();
 	c.dump_config(config);
 	REQUIRE(config.size() == 3);
 	REQUIRE(equivalent());
+
+	expected.emplace("color hint-keys-delimiter color5 default dim");
+	c.handle_action("color", {"hint-keys-delimiter", "color5", "default", "dim"});
+	config.clear();
+	c.dump_config(config);
+	REQUIRE(config.size() == 4);
+	REQUIRE(equivalent());
+
+	expected.emplace("color hint-separator color7 color8");
+	c.handle_action("color", {"hint-separator", "color7", "color8"});
+	config.clear();
+	c.dump_config(config);
+	REQUIRE(config.size() == 5);
+	REQUIRE(equivalent());
+
+}
+
+TEST_CASE("If no colors were specified for the "
+	"`title`/`hint-key`/`hint-keys-delimiter`/`hint-separator`/`hint-description` elements, "
+	"then use colors from the `info` element (if any)",
+	"[ColorManager]")
+{
+	const std::vector<std::string> elements {
+		"title", "hint-key", "hint-keys-delimiter", "hint-separator", "hint-description"};
+
+	for (const auto& element : elements) {
+		DYNAMIC_SECTION("element: " << element) {
+			ColorManager c;
+			StylesCollector collector;
+
+			SECTION("Element's style can be changed as usual") {
+				c.handle_action("color", {element, "green", "default", "underline"});
+
+				c.apply_colors(collector.setter());
+
+				REQUIRE(collector.styles_count() >= 1);
+				REQUIRE(collector.style(element) == "fg=green,attr=underline");
+			}
+
+			SECTION("Element and `info` don't interfere with each other") {
+				SECTION("Element's style is set before `info`") {
+					c.handle_action("color", {element, "blue", "black"});
+					c.handle_action("color", {"info", "green", "yellow", "bold"});
+				}
+
+				SECTION("Element's style is set after `info`") {
+					c.handle_action("color", {"info", "green", "yellow", "bold"});
+					c.handle_action("color", {element, "blue", "black"});
+				}
+
+				c.apply_colors(collector.setter());
+
+				REQUIRE(collector.styles_count() >= 2);
+				REQUIRE(collector.style(element) == "fg=blue,bg=black");
+				REQUIRE(collector.style("info") == "fg=green,bg=yellow,attr=bold");
+			}
+
+			SECTION("Element inherits the `info` style when available") {
+				c.handle_action("color", {"info", "red", "magenta", "reverse"});
+
+				c.apply_colors(collector.setter());
+
+				REQUIRE(collector.styles_count() >= 2);
+				REQUIRE(collector.style(element) == "fg=red,bg=magenta,attr=reverse");
+				REQUIRE(collector.style("info") == "fg=red,bg=magenta,attr=reverse");
+			}
+
+			SECTION("Element has no style if there is no style for `info` either") {
+				c.handle_action("color", {"listnormal", "black", "white"});
+
+				c.apply_colors(collector.setter());
+
+				REQUIRE(collector.styles_count() >= 1);
+				REQUIRE(collector.style("listnormal") == "fg=black,bg=white");
+
+				REQUIRE(collector.style(element) == "");
+				REQUIRE(collector.style("info") == "");
+			}
+		}
+	}
 }